#4 myPage components 분리
